Added pre-processing of a given file to remove rows with missing data.
The preprocessing.py program loops through a .csv file and removes rows of data that have missing data values for four attributes of interest for clustering algorithm tests. The result is the creation of a new, smaller data set that excludes the offending rows.

Note: There are still many rows with missing data from other attribute columns. If data from those columns become needed, the code from this program would need to be physically altered as the specific columns of interest are currently hard coded.