Changed the uploaded data set to be parsed as a 3D K-Means scatter plot.
Previously the data set uploaded was parsed and displayed as a DataTable. Now it is being parsed and displayed as a 3D scatter plot.  The parsed data is run through the scikit-learn KMeans algorithm as a first test. There is not other user input at this time. Data attributes for x, y, and z axes are hard coded for the purposes of this test.