Stacked two identical scatter plot graphs horizontally for future comparison feature.
The two scatter plots take the same data set that was uploaded by the user and call the same file parsing to K-Means function and display the graphs side-by-side. This is a precursor to the intended function (with the uploaded file the two app callback functions will parse the file and run separate algorithms).