Added in logic to allow axis selection dropdown menus to appear only after an appacptable file has been uploaded and parsed.
The axis choice dropdown menus (for x, y, and z axes) are made in the first callback function that takes in a data set from an uploaded file, parses it to a Pandas data frame, and collects the numerical-based attribute/axis options from that data frame. These options are stored in a global variable that can be accessed by the dropdown options property when they are made within that same first callback function (returned as children output to a Div container within the app layout).

The K-Means graphs were omitted during this integration phase for simplicity of testing the functionality to the dropdown menus.