The algorithm dropdown menu selection now triggers the K-Means figure function.
The update_input_dropdown callback function waits until an acceptable file has been uploaded, parses it (collecting the axes' options), stores the data frame in a global variable and returns the data frame. This data frame is called in the update_output_graph1 callback function where it is currently used to make a 2D and 3D K-Means scatter plot (used as a parameter to the appropriate K-Means functions).

The other algorithm options from the drop down only produce a placeholder statement. The selection of various algorithms successfully updates the respective outputs in the web app.