Added the 2D vs 3D graph selection dropdown.
The program now offers algorithm selection, choice of 2D or 3D graph for certain algorithms, and x, y, and z axis dropdowns. Was not able to successfully separate these components into individual callbacks.

Program can output correct 2D (in the case of K-Means or GMM) or 3D (in the case of K-Means) graphs when appropriate categories are selected.  NOTE: Need to add in output statements for the user, stating that if they selected 3D for an algorithm that does not offer 3D that this is the case and to not expect a 3D graph.