Updated the parsing function to include a check for enough usable data columns to make a at least a 2D graph.
This function now returns an int value and is stored in a new variable in the function call in the update_input_dropdowns callback function.  This is used as a means to allow a user to proceed to the interactive selectors, or to provide a warning message saying there isn't enough data (and no interactive selectors are made visible).

This is the first part of the intended upgraded pre-processing feature.